+++ Backend/server.js
@@ -2,8 +2,8 @@ import express from "express";
 import { PORT, mongoDBUrl } from "./config.js";
 import cors from "cors";
 import mongoose from "mongoose";
-import workoutRoutes from "./Routes/workouts.js";
-import usersRoutes from "./Routes/users.js";
+import workoutRoutes from "./Routes/workoutRoute.js";
+import usersRoutes from "./Routes/userRoute.js";
 
 const app = express();
